MYAUT-98-dividing-listing-service-into-2-parts-profile-service-and-listing-service-as-controllers done
