+++ myauto/src/main/java/az/code/myauto/services/interfaces/ImageService.java
@@ -9,10 +9,16 @@ import org.springframework.web.multipart.MultipartFile;
 import java.util.List;
 
 public interface ImageService {
+
     List<ImageDTO> getImagesByListingId(Long id) throws ThumbnailNotFoundException;
+
     ImageDTO getImageByListingId(Long listingId, Long id) throws ThumbnailNotFoundException;
+
     String uploadImageToFireBase(UserDTO user, MultipartFile multipartFile);
+
     ImageDTO addImageToListing(Long id, UserDTO user, String url);
+
     Image isImageExist(Long listingId, Long id) throws ThumbnailNotFoundException;
+
     void deleteImage(UserDTO user, Long listingId, Long id) throws ThumbnailNotFoundException;
 }
