+++ myauto/src/main/java/az/code/myauto/services/interfaces/ListingService.java
@@ -11,29 +11,10 @@ import java.util.List;
 
 public interface ListingService {
 
-    ListingGetDTO create(ListingCreationDTO listing, UserDTO user) throws FreeListingAlreadyPostedException;
-
-    ListingGetDTO update(long id, ListingCreationDTO listing, UserDTO user) throws ListingNotFoundException;
-
-    void delete(long id, UserDTO user) throws ListingNotFoundException;
-
-    ListingGetDTO makeVip(long id, UserDTO user) throws ListingNotFoundException, TransactionIncorrectAmountException, TransactionInsufficientFundsException;
-
-    ListingGetDTO makePaid(long id, UserDTO user);
-
-    ListingGetDTO setNewThumbnail(long id, UserDTO user, ImageDTO imageDTO);
-
-    List<ListingListDTO> getUserListings(Integer pageNo, Integer pageSize, String sortBy, UserDTO user);
-
-    ListingGetDTO getUserListingById(long id, UserDTO user) throws ListingNotFoundException;
-
-
     ListingGetDTO getListingById(long id) throws ListingNotFoundException;
 
     List<ListingListDTO> getListings(Integer pageNo, Integer pageSize, String sortBy);
 
     List<ListingListDTO> getVIPListings(Integer pageNo, Integer pageSize, String sortBy);
-
-    Listing isListingExist(long id, UserDTO user) throws ListingNotFoundException;
 
 }
